requirements.txt
amqp==2.2.2
anyjson==0.3.3
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
cffi==1.11.0
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
decorator==4.1.2
dj-database-url==0.4.2
Django==1.11.6
django-address==0.1.5
django-celery==3.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-dynamic-scraper==0.13.0
django-filter==1.0.4
django-kombu==0.9.4
django-multiselectfield==0.1.7
django-widget-tweaks==1.4.1
djangorestframework==3.7.1
enum34==1.1.6
future==0.15.2
geoip2==2.5.0
googlemaps==2.4.6
gunicorn==19.7.1
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
jsonpath-rw==1.4.0
kombu==4.1.0
librabbitmq==1.6.1
lxml==3.8.0
maxminddb==1.3.0
parsel==1.2.0
Pillow==3.4.2
ply==3.10
psycopg2==2.7.3.1
pyasn1==0.3.5
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.3.0
python-dateutil==2.6.1
pytz==2017.2
queuelib==1.4.2
redis==2.10.6
requests==2.11.1
Scrapy==1.4.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
vine==1.1.4
w3lib==1.18.0
whitenoise==3.3.0
zope.interface==4.4.2
